requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
python-dateutil==2.8.2
